% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_10.R
\name{log_10}
\alias{log_10}
\title{Log (base 10) of a vector.}
\usage{
log_10(x)
}
\arguments{
\item{x}{The vector and power to log(10).}
}
\value{
A vector that is the exponent to raise 10 to get the power, x \code{x}.
}
\description{
This function just takes the log (base 10) of a vector.
}
\examples{
log_10(1:10)
log_10(-5)
}
