% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Raise the vector x to the exponent y.}
\usage{
pow(x, y = 2)
}
\arguments{
\item{x}{The vector to be raised to the exponent y.}

\item{y}{The exponent.}
}
\value{
A vector that is raised to the exponent y \code{x}.
}
\description{
This function raises x to the exponent y.
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{pow} function is quite simple.
     \item There's nothing else to say about \code{pow}.
}
}
\examples{
pow(1:10)
pow(-5)
}
