% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers-package.r
\docType{package}
\name{Powers}
\alias{Powers}
\alias{Powers-package}
\title{Powers.}
\description{
This package was created for a STAT547 assignment to practice developing packages in R.
It contains two functions, log_10 and cube. These can be used to calculate the log base ten of and cubed values of a vector.
There is more detail about this package in the vignette file.
}
