% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_10.R
\name{log_10}
\alias{log_10}
\title{Log (base 10) of a vector.}
\usage{
log_10(x)
}
\arguments{
\item{x}{The vector to log_10.}
}
\value{
A vector that is the exponent to raise 10 to get the power, x \code{x}.
}
\description{
This function takes the log (base 10) of a vector.
}
\details{
Logarithmic functions are the inverses of exponential functions.
}
\examples{
log_10(1:10)
log_10(-5)
}
